+++ docs-markdown/src/controllers/alert-controller.ts
@@ -1,10 +1,9 @@
 "use strict";
 
 import * as vscode from "vscode";
-import { AlertTags } from "../constants/alert-tags";
 import { insertContentToEditor, isMarkdownFileCheck, noActiveEditorMessage } from "../helper/common";
-import { reporter } from "../telemetry/telemetry";
 import { format } from "../helper/format";
+import { reporter } from "../telemetry/telemetry";
 
 const telemetryCommand: string = "insertAlert";
 
@@ -65,29 +64,5 @@ export function insertAlert() {
                 }
             }
         });
-    }
-}
-
-/**
- * Returns the alert type
- * @param {string} content - the string content
- * @return {AlertType} - the type of alert i.e. Note, Warning, Important, Tip
- */
-export function getAlertType(content: string) {
-    return AlertTags.findIndex((tag) => content.startsWith(tag));
-}
-
-/**
- * Checks if the string input is a valid alert
- * @param {string} content - the string content
- * @return {boolean} - true/false the content is an alert
- */
-export function isAlert(content: string) {
-    // Check if the content starts with an alert tag and if all paragraphs contain the ">" formatter
-    if ((AlertTags.some((tag) => content.startsWith(tag))) &&
-        (content.split("\n").every((line) => line.startsWith(">")))) {
-        return true;
-    } else {
-        return false;
     }
 }
